Reorganize settings into drill-down categories
The settings screen had grown to a single 3,061-line file rendering 13
sections / ~54 controls in one endless scroll — hard for users to navigate
and hard to maintain. Convert it to a drill-down (master/detail) layout and
split the monolith into one file per category.

The landing page now shows the operator card pinned at top, the GPS
auto-grid toggle, and six tappable categories. Each opens a focused detail
screen with a back chevron. Drill-down is driven by internal state (no
NavHost) since Settings is hosted as a plain tab swap in FT8USApp; hardware
back pops detail->landing and propagates (app exit) on the landing.

State is now scoped per category (read from GeneralVariables, written via
writeConfig), which also fixes the old "one toggle recomposes the whole
3k-line tree" problem. Shared helpers and generic dialogs live in
SettingsCommon.kt; category-specific dialogs live with their category.

New files: SettingsCommon, RadioAudioSettings, TransmissionSettings,
DecodeFilterSettings, LoggingSettings, AdvancedSettings, AboutSettings.
SettingsScreen.kt is now the slim host + landing + operator dialog (429
lines, down from 3,061). Adds six localized category-label strings across
all seven locales.
