+++ ft8cn/app/src/main/kotlin/radio/ks3ckc/ft8us/ui/settings/DecodeFilterSettings.kt
@@ -1,0 +1,389 @@
+package radio.ks3ckc.ft8us.ui.settings
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.bg7yoz.ft8cn.GeneralVariables
+import com.bg7yoz.ft8cn.MainViewModel
+import com.bg7yoz.ft8cn.R
+import radio.ks3ckc.ft8us.ui.components.GlassCard
+import radio.ks3ckc.ft8us.ui.components.SettingsRow
+
+/**
+ * Decode-list settings: highlight rules, callsign blocklist, display filters,
+ * and needed-DX alerts.
+ */
+@Composable
+fun DecodeFilterSettings(
+    mainViewModel: MainViewModel,
+    onBack: () -> Unit,
+) {
+    // Decode-list highlight toggles
+    var highlightNewDxcc by remember { mutableStateOf(GeneralVariables.highlightNewDxcc) }
+    var highlightNewGrid by remember { mutableStateOf(GeneralVariables.highlightNewGrid) }
+    var highlightNewBand by remember { mutableStateOf(GeneralVariables.highlightNewBand) }
+    var highlightWorked by remember { mutableStateOf(GeneralVariables.highlightWorked) }
+    var highlightPota by remember { mutableStateOf(GeneralVariables.highlightPota) }
+    var distanceInMiles by remember { mutableStateOf(GeneralVariables.distanceInMiles) }
+
+    // Callsign blocklist (comma-separated entries) + decode display filters
+    var blockedExact by remember { mutableStateOf(GeneralVariables.getBlockedExactCallsigns()) }
+    var blockedPrefixes by remember { mutableStateOf(GeneralVariables.getExcludeCallsigns()) }
+    var blockedKeywords by remember { mutableStateOf(GeneralVariables.getBlockedKeywords()) }
+    var filterShowOnlyCQ by remember { mutableStateOf(GeneralVariables.filterShowOnlyCQ) }
+    var filterDxOnly by remember { mutableStateOf(GeneralVariables.filterDxOnly) }
+    var filterNeededOnly by remember { mutableStateOf(GeneralVariables.filterNeededOnly) }
+    var filterByContinent by remember { mutableStateOf(GeneralVariables.filterByContinent) }
+    var filterContinent by remember { mutableStateOf(GeneralVariables.filterContinent) }
+    var respectDirectionalCQ by remember { mutableStateOf(GeneralVariables.respectDirectionalCQ) }
+    var filterDirectionalCQ by remember { mutableStateOf(GeneralVariables.filterDirectionalCQ) }
+    var alertNewDxcc by remember { mutableStateOf(GeneralVariables.alertNewDxcc) }
+    var alertNewState by remember { mutableStateOf(GeneralVariables.alertNewState) }
+
+    // Continent codes (stored on the message) and their display names, parallel lists.
+    val continentCodes = listOf("NA", "SA", "EU", "AF", "AS", "OC", "AN")
+    val continentNames = listOf(
+        stringResource(R.string.continent_na),
+        stringResource(R.string.continent_sa),
+        stringResource(R.string.continent_eu),
+        stringResource(R.string.continent_af),
+        stringResource(R.string.continent_as),
+        stringResource(R.string.continent_oc),
+        stringResource(R.string.continent_an),
+    )
+
+    var showBlockExactDialog by remember { mutableStateOf(false) }
+    var showBlockPrefixDialog by remember { mutableStateOf(false) }
+    var showBlockKeywordDialog by remember { mutableStateOf(false) }
+    var showContinentPicker by remember { mutableStateOf(false) }
+
+    // -- Blocklist: exact whole-call dialog --
+    if (showBlockExactDialog) {
+        TextListDialog(
+            title = stringResource(R.string.settings_block_exact_title),
+            description = stringResource(R.string.settings_block_exact_desc),
+            initialValue = blockedExact,
+            onDismiss = { showBlockExactDialog = false },
+            onSave = { text ->
+                GeneralVariables.addBlockedExactCallsigns(text)
+                blockedExact = GeneralVariables.getBlockedExactCallsigns()
+                mainViewModel.databaseOpr.writeConfig("blockedExactCallsigns", blockedExact, null)
+                showBlockExactDialog = false
+            },
+        )
+    }
+
+    // -- Blocklist: prefix dialog (legacy excludedCallsigns key) --
+    if (showBlockPrefixDialog) {
+        TextListDialog(
+            title = stringResource(R.string.settings_block_prefix_title),
+            description = stringResource(R.string.settings_block_prefix_desc),
+            initialValue = blockedPrefixes,
+            onDismiss = { showBlockPrefixDialog = false },
+            onSave = { text ->
+                GeneralVariables.addExcludedCallsigns(text)
+                blockedPrefixes = GeneralVariables.getExcludeCallsigns()
+                mainViewModel.databaseOpr.writeConfig("excludedCallsigns", blockedPrefixes, null)
+                showBlockPrefixDialog = false
+            },
+        )
+    }
+
+    // -- Blocklist: keyword dialog --
+    if (showBlockKeywordDialog) {
+        TextListDialog(
+            title = stringResource(R.string.settings_block_keyword_title),
+            description = stringResource(R.string.settings_block_keyword_desc),
+            initialValue = blockedKeywords,
+            onDismiss = { showBlockKeywordDialog = false },
+            onSave = { text ->
+                GeneralVariables.addBlockedKeywords(text)
+                blockedKeywords = GeneralVariables.getBlockedKeywords()
+                mainViewModel.databaseOpr.writeConfig("blockedKeywords", blockedKeywords, null)
+                showBlockKeywordDialog = false
+            },
+        )
+    }
+
+    // -- Decode filter: continent picker --
+    if (showContinentPicker) {
+        val currentIndex = continentCodes.indexOf(filterContinent).coerceAtLeast(0)
+        ListPickerDialog(
+            title = stringResource(R.string.settings_filter_continent_title),
+            items = continentNames,
+            selectedIndex = currentIndex,
+            onDismiss = { showContinentPicker = false },
+            onSelect = { index ->
+                showContinentPicker = false
+                val code = continentCodes[index]
+                filterContinent = code
+                GeneralVariables.filterContinent = code
+                mainViewModel.databaseOpr.writeConfig("filterContinent", code, null)
+            },
+        )
+    }
+
+    SettingsDetailScaffold(
+        title = stringResource(R.string.settings_cat_decode_filters),
+        onBack = onBack,
+    ) {
+        // =====================================================================
+        // DECODE HIGHLIGHTS
+        // =====================================================================
+        SettingsSection(title = stringResource(R.string.settings_section_decode_highlights)) {
+            GlassCard(modifier = Modifier.fillMaxWidth()) {
+                Column {
+                    SettingsRow(
+                        label = stringResource(R.string.settings_highlight_new_dxcc),
+                        description = stringResource(R.string.settings_highlight_new_dxcc_desc),
+                        toggle = highlightNewDxcc,
+                        onToggleChange = { checked ->
+                            highlightNewDxcc = checked
+                            GeneralVariables.highlightNewDxcc = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "highlightNewDxcc", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_highlight_new_grid),
+                        description = stringResource(R.string.settings_highlight_new_grid_desc),
+                        toggle = highlightNewGrid,
+                        onToggleChange = { checked ->
+                            highlightNewGrid = checked
+                            GeneralVariables.highlightNewGrid = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "highlightNewGrid", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_highlight_new_band),
+                        description = stringResource(R.string.settings_highlight_new_band_desc),
+                        toggle = highlightNewBand,
+                        onToggleChange = { checked ->
+                            highlightNewBand = checked
+                            GeneralVariables.highlightNewBand = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "highlightNewBand", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_highlight_pota),
+                        description = stringResource(R.string.settings_highlight_pota_desc),
+                        toggle = highlightPota,
+                        onToggleChange = { checked ->
+                            highlightPota = checked
+                            GeneralVariables.highlightPota = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "highlightPota", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_highlight_worked),
+                        description = stringResource(R.string.settings_highlight_worked_desc),
+                        toggle = highlightWorked,
+                        onToggleChange = { checked ->
+                            highlightWorked = checked
+                            GeneralVariables.highlightWorked = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "highlightWorked", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_distance_unit),
+                        description = stringResource(R.string.settings_distance_unit_desc),
+                        toggle = distanceInMiles,
+                        onToggleChange = { checked ->
+                            distanceInMiles = checked
+                            GeneralVariables.distanceInMiles = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "distanceInMiles", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                }
+            }
+        }
+
+        // =====================================================================
+        // CALLSIGN BLOCKLIST
+        // =====================================================================
+        SettingsSection(title = stringResource(R.string.settings_section_callsign_blocklist)) {
+            GlassCard(modifier = Modifier.fillMaxWidth()) {
+                Column {
+                    SettingsRow(
+                        label = stringResource(R.string.settings_exact_callsigns),
+                        description = stringResource(R.string.settings_exact_callsigns_desc),
+                        value = blockedExact.ifBlank { stringResource(R.string.common_none) },
+                        showChevron = true,
+                        onClick = { showBlockExactDialog = true },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_prefixes),
+                        description = stringResource(R.string.settings_prefixes_desc),
+                        value = blockedPrefixes.ifBlank { stringResource(R.string.common_none) },
+                        showChevron = true,
+                        onClick = { showBlockPrefixDialog = true },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_keywords),
+                        description = stringResource(R.string.settings_keywords_desc),
+                        value = blockedKeywords.ifBlank { stringResource(R.string.common_none) },
+                        showChevron = true,
+                        onClick = { showBlockKeywordDialog = true },
+                    )
+                }
+            }
+        }
+
+        // =====================================================================
+        // DECODE FILTERS
+        // =====================================================================
+        SettingsSection(title = stringResource(R.string.settings_section_decode_filters)) {
+            GlassCard(modifier = Modifier.fillMaxWidth()) {
+                Column {
+                    SettingsRow(
+                        label = stringResource(R.string.settings_show_only_cq),
+                        description = stringResource(R.string.settings_show_only_cq_desc),
+                        toggle = filterShowOnlyCQ,
+                        onToggleChange = { checked ->
+                            filterShowOnlyCQ = checked
+                            GeneralVariables.filterShowOnlyCQ = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "filterShowOnlyCQ", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_dx_only),
+                        description = stringResource(R.string.settings_dx_only_desc),
+                        toggle = filterDxOnly,
+                        onToggleChange = { checked ->
+                            filterDxOnly = checked
+                            GeneralVariables.filterDxOnly = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "filterDxOnly", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_needed_only),
+                        description = stringResource(R.string.settings_needed_only_desc),
+                        toggle = filterNeededOnly,
+                        onToggleChange = { checked ->
+                            filterNeededOnly = checked
+                            GeneralVariables.filterNeededOnly = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "filterNeededOnly", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_filter_by_continent),
+                        description = stringResource(R.string.settings_filter_by_continent_desc),
+                        toggle = filterByContinent,
+                        onToggleChange = { checked ->
+                            filterByContinent = checked
+                            GeneralVariables.filterByContinent = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "filterByContinent", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    if (filterByContinent) {
+                        SectionDivider()
+                        SettingsRow(
+                            label = stringResource(R.string.settings_continent),
+                            value = continentNames.getOrElse(
+                                continentCodes.indexOf(filterContinent),
+                            ) { filterContinent },
+                            showChevron = true,
+                            onClick = { showContinentPicker = true },
+                        )
+                    }
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_skip_directional_cq),
+                        description = stringResource(R.string.settings_skip_directional_cq_desc),
+                        toggle = respectDirectionalCQ,
+                        onToggleChange = { checked ->
+                            respectDirectionalCQ = checked
+                            GeneralVariables.respectDirectionalCQ = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "respectDirectionalCQ", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_hide_directional_cq),
+                        description = stringResource(R.string.settings_hide_directional_cq_desc),
+                        toggle = filterDirectionalCQ,
+                        onToggleChange = { checked ->
+                            filterDirectionalCQ = checked
+                            GeneralVariables.filterDirectionalCQ = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "filterDirectionalCQ", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                }
+            }
+        }
+
+        // =====================================================================
+        // NEEDED-DX ALERTS
+        // =====================================================================
+        SettingsSection(title = stringResource(R.string.settings_section_needed_dx_alerts)) {
+            GlassCard(modifier = Modifier.fillMaxWidth()) {
+                Column {
+                    SettingsRow(
+                        label = stringResource(R.string.settings_alert_new_dxcc),
+                        description = stringResource(R.string.settings_alert_new_dxcc_desc),
+                        toggle = alertNewDxcc,
+                        onToggleChange = { checked ->
+                            alertNewDxcc = checked
+                            GeneralVariables.alertNewDxcc = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "alertNewDxcc", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                    SectionDivider()
+                    SettingsRow(
+                        label = stringResource(R.string.settings_alert_new_state),
+                        description = stringResource(R.string.settings_alert_new_state_desc),
+                        toggle = alertNewState,
+                        onToggleChange = { checked ->
+                            alertNewState = checked
+                            GeneralVariables.alertNewState = checked
+                            mainViewModel.databaseOpr.writeConfig(
+                                "alertNewState", if (checked) "1" else "0", null,
+                            )
+                        },
+                    )
+                }
+            }
+        }
+    }
+}
