+++ ft8cn/app/src/main/res/values-ja/strings_compose.xml
@@ -291,6 +291,14 @@
     <string name="settings_section_about">情報</string>
     <string name="settings_section_language">言語</string>
 
+    <!-- ===== Settings: drill-down category labels ===== -->
+    <string name="settings_cat_radio_audio">無線とオーディオ</string>
+    <string name="settings_cat_transmission">送信</string>
+    <string name="settings_cat_decode_filters">デコードとフィルター</string>
+    <string name="settings_cat_logging">ログと実績</string>
+    <string name="settings_cat_advanced">詳細設定</string>
+    <string name="settings_cat_about">情報</string>
+
     <!-- ===== Settings: operator identity ===== -->
     <string name="settings_auto_update_grid">グリッドを自動更新 (GPS)</string>
     <string name="settings_auto_update_grid_desc">デバイスの GPS を使って Maidenhead グリッドを最新に保ちます</string>
